--- LearningProject/src/java/Controller/GoogleLoginController/LoginGoogleController.java
@@ -1,48 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/JSP_Servlet/Servlet.java to edit this template
- */
-package Controller.GoogleLoginController;
-
-import Model.GooglePojo;
-import Model.User;
-import dal.UserDAO;
-import jakarta.servlet.RequestDispatcher;
-import java.io.IOException;
-import java.io.PrintWriter;
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import utils.GoogleUtils;
-
-/**
- *
- * @author NamDepTraiVL
- */
-@WebServlet(name = "LoginGoogleController", urlPatterns = {"/login-google"})
-public class LoginGoogleController extends HttpServlet {
-
-    private static final long serialVersionUID = 1L;
-
-        protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        String code = request.getParameter("code");
-        if (code == null || code.isEmpty()) {
-            request.getRequestDispatcher("Login.jsp").forward(request, response);
-        } else {
-            String accessToken = GoogleUtils.getToken(code);
-            GooglePojo googlePojo = GoogleUtils.getUserInfo(accessToken);
-            User user = new User(googlePojo.getGiven_name(), googlePojo.getFamily_name(), googlePojo.getEmail(), googlePojo.getId());
-            UserDAO uDao = new UserDAO();
-            uDao.addUser(user);            
-        }
-    }
-
-    protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws ServletException, IOException {
-        doGet(request, response);
-    }
-
-}
