+++ LearningProject/web/WEB-INF/web.xml
@@ -9,6 +9,14 @@
         <servlet-class>Controller.ResetController.ForgotPass</servlet-class>
     </servlet>
     <servlet>
+        <servlet-name>GoogleLoginController</servlet-name>
+        <servlet-class>Controller.GoogleLoginController.GoogleLoginController</servlet-class>
+    </servlet>
+    <servlet-mapping>
+        <servlet-name>GoogleLoginController</servlet-name>
+        <url-pattern>/login-google</url-pattern>
+    </servlet-mapping>
+    <servlet>
         <servlet-name>AvatarUpdate</servlet-name>
         <servlet-class>Controller.AvatarUpdate</servlet-class>
     </servlet>
